aug 17
Final commit before live, all commits after this will be to modify the
live site and update the features. I wil need to update to django 1.10
and alter some of the code to make the login and logout work in 1.10 as
this breaks a few things. I will also look at add features and generally
improve the site overall