Aug 19 V1.5
working on the first major update for the site first commit for that.
Removed all the outdated componants now: Login, Logout, Blog create and
Blog update, still have an issue with saving images correctly but the
class is getting closer. I have worked on some CSS also to make the blog
style a little better